[f] OUT-77 - Check for reply in 10 minutes timeframe after first message

+++ supabase/functions/user-actions/handlers/broadcast-reply-handlers.ts
@@ -9,27 +9,21 @@ export const deletePendingSecondBroadcastMessage = async (
   requestBody: RequestBody,
 ) => {
   const requestMessage = requestBody.message!
-  const deliveredAtUnixTimestamp = requestMessage.delivered_at * 1000
-  const deliveredDate = new Date(deliveredAtUnixTimestamp)
-  deliveredDate.setHours(0, 0, 0, 0)
-  const nextDay = new Date(deliveredDate)
-  nextDay.setDate(deliveredDate.getDate() + 1)
-  const deletedMessages = await db.delete(outgoingMessages)
-    .where(
-      and(
-        // gte(outgoingMessages.createdAt, deliveredDate.getTime() / 1000), // TODO: get delay from broadcast
-        // lt(outgoingMessages.createdAt, nextDay.getTime() / 1000),
-        eq(outgoingMessages.isSecond, true),
-        eq(
-          outgoingMessages.recipientPhoneNumber,
-          requestMessage.from_field.id,
+  const firstMessage = await db.select().from(outgoingMessages).where(
+    and(
+      eq(outgoingMessages.isSecond, false),
+      eq(outgoingMessages.recipientPhoneNumber, requestMessage.from_field.id),
+      eq(outgoingMessages.processed, false),
+    ),
+  )
+  if (firstMessage.length === 0) {
+    // First message was sent
+    await db.delete(outgoingMessages)
+      .where(
+        and(
+          eq(outgoingMessages.isSecond, true),
+          eq(outgoingMessages.recipientPhoneNumber, requestMessage.from_field.id),
         ),
-      ),
-    )
-    .returning()
-
-  console.log(
-    'Deleted pending second message rows from OutgoingMessages ',
-    deletedMessages,
-  ) // Log any affected rows
+      )
+  }
 }
